# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HappyDrive API
  version: 1.1.0
  description: |
    日本語iOS・企業ポータル・運営Web共通契約。全書込に監査イベントを記録する。
    v1.0（docs/spec/openapi.v1.0-original.yaml）からの変更は後方互換の追加のみ。変更履歴は docs/CONTRACT_CHANGELOG.md。
    - 時刻はUTC（ISO 8601）で送受信し、表示はJST。金額は整数JPY。
    - エラーは常に Error スキーマ（message は日本語）。
    - 重要な書込は Idempotency-Key 必須。同一キー・同一内容の再送は同じ応答を返し、内容が異なれば 422 idempotency_mismatch。
servers:
  - url: https://api.happydrive.jp/v1
  - url: http://localhost:8080/v1
security:
  - bearerAuth: []
...
paths:
...
  /earnings:
    get:
      tags: [Earnings]
      operationId: getEarnings
      parameters: [{name: month, in: query, schema: {type: string, pattern: '^[0-9]{4}-[0-9]{2}$'}}]
      responses:
        default: {$ref: '#/components/responses/Default'}
        '200': {description: 報酬台帳, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/EarningEntry'}}}}}
  /earnings/summary:
    get:
      tags: [Earnings]
      operationId: getEarningsSummary
      parameters: [{name: month, in: query, schema: {type: string, pattern: '^[0-9]{4}-[0-9]{2}$'}}]
      responses:
        default: {$ref: '#/components/responses/Default'}
        '200': {description: 月次集計, content: {application/json: {schema: {$ref: '#/components/schemas/EarningsSummary'}}}}
  /payouts:
    get:
      tags: [Earnings]
      operationId: listMyPayouts
      responses:
        default: {$ref: '#/components/responses/Default'}
        '200': {description: 振込履歴, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Payout'}}}}}

  # ---------------------------------------------------------------- Learning
...
  /admin/payouts:
    get:
      tags: [Admin]
      operationId: adminListPayouts
      parameters: [{name: status, in: query, schema: {type: string, enum: [requested, processing, paid, failed]}}]
      responses:
        default: {$ref: '#/components/responses/Default'}
        '200': {description: 振込, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Payout'}}}}}
  /admin/payouts/batches:
    post:
      tags: [Admin]
      operationId: adminCreatePayoutBatch
      description: payable の報酬を利用者ごとに集計して振込要求を作成し、決済事業者へ送信。二重振込防止キーを利用。決済事業者未契約の本番構成では 503。
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [cutoffDate], properties: {cutoffDate: {type: string, format: date}}}}}}
      responses:
        default: {$ref: '#/components/responses/Default'}
        '201': {description: 作成, content: {application/json: {schema: {$ref: '#/components/schemas/PayoutBatch'}}}}
        '503': {$ref: '#/components/responses/Unavailable'}
  /admin/payouts/{payoutId}/retry:
    post:
      tags: [Admin]
      operationId: adminRetryPayout
      parameters: [{name: payoutId, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        default: {$ref: '#/components/responses/Default'}
        '200': {description: 再試行, content: {application/json: {schema: {$ref: '#/components/schemas/Payout'}}}}
        '409': {$ref: '#/components/responses/Conflict'}
  /admin/reconciliation:
    get:
      tags: [Admin]
      operationId: adminReconciliation
      description: 月次照合。受諾スナップショット合計・台帳合計・振込合計・決済事業者の確認済み額の差異を返す。
      parameters: [{name: month, in: query, required: true, schema: {type: string, pattern: '^[0-9]{4}-[0-9]{2}$'}}]
      responses:
        default: {$ref: '#/components/responses/Default'}
        '200': {description: 照合結果, content: {application/json: {schema: {$ref: '#/components/schemas/Reconciliation'}}}}
...
  /webhooks/payments/{provider}:
    post:
      tags: [Webhooks]
      operationId: paymentWebhook
      security: []
      description: 決済事業者Webhook。署名検証・イベントID一意・順不同耐性。入金/振込の成立はWebhookのみでなく事業者APIで照合する。
      parameters: [{name: provider, in: path, required: true, schema: {type: string}}]
      requestBody: {required: true, content: {application/json: {schema: {type: object}}}}
      responses:
        default: {$ref: '#/components/responses/Default'}
        '200': {description: 受理（重複も200）}
        '401': {$ref: '#/components/responses/Unauthorized'}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
  responses:
    Unauthorized: {description: 認証が必要, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    Conflict: {description: 状態・残枠・重複要求の競合, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    Unavailable: {description: 依存先停止または未契約, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Default: {description: その他のエラー（401 認証 / 403 権限 / 404 不存在・他組織 / 422 入力 / 429 回数制限 / 5xx）。本文は常に Error, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: {type: string, description: '機械判定用コード（例 capacity_full / not_eligible / invalid_state）'}
        message: {type: string, description: 利用者に表示できる日本語}
        requestId: {type: string}
        details: {type: object, additionalProperties: true}
...
    ContractType:
      type: string
      enum: [employment, contractor, other_legal_review]
      description: other_legal_review（その他・法務審査中）は公開不可。
...
    EarningEntry:
      type: object
      required: [assignmentId, amountYen, state]
      properties:
        id: {type: string, format: uuid}
        assignmentId: {type: string, format: uuid}
        jobTitle: {type: string}
        organizationName: {type: string}
        contractType: {$ref: '#/components/schemas/ContractType'}
        amountYen: {type: integer}
        state: {type: string, enum: [estimated, pending, payable, paid, failed, reversed]}
        scheduledPayoutDate: {type: string, format: date}
        paidAt: {type: string, format: date-time}
        workDate: {type: string, format: date}
        breakdown: {type: array, items: {type: object, required: [type, amountYen], properties: {type: {type: string}, amountYen: {type: integer}, note: {type: string}}}}
    EarningsSummary:
      type: object
      required: [month, totalYen, paidYen, payableYen, pendingYen, estimatedYen, completedCount]
      properties: {month: {type: string}, totalYen: {type: integer}, paidYen: {type: integer}, payableYen: {type: integer}, pendingYen: {type: integer}, estimatedYen: {type: integer}, completedCount: {type: integer}, employmentYen: {type: integer}, contractorYen: {type: integer}}
    Payout:
      type: object
      required: [id, amountYen, status, createdAt]
      properties: {id: {type: string, format: uuid}, workerId: {type: string, format: uuid}, workerDisplayName: {type: string}, amountYen: {type: integer}, status: {type: string, enum: [requested, processing, paid, failed]}, failureReason: {type: string}, providerReference: {type: string}, attemptCount: {type: integer}, scheduledDate: {type: string, format: date}, paidAt: {type: string, format: date-time}, createdAt: {type: string, format: date-time}}
    PayoutBatch:
      type: object
      required: [id, cutoffDate, payoutCount, totalYen]
      properties: {id: {type: string, format: uuid}, cutoffDate: {type: string, format: date}, payoutCount: {type: integer}, totalYen: {type: integer}, payouts: {type: array, items: {$ref: '#/components/schemas/Payout'}}}
...
    Reconciliation:
      type: object
      required: [month, acceptedSnapshotYen, ledgerEarnedYen, ledgerReversedYen, payoutsPaidYen, providerConfirmedYen, discrepancies]
      properties:
        month: {type: string}
        acceptedSnapshotYen: {type: integer}
        ledgerEarnedYen: {type: integer}
        ledgerReversedYen: {type: integer}
        payoutsPaidYen: {type: integer}
        providerConfirmedYen: {type: integer}
        failedPayoutCount: {type: integer}
        discrepancies: {type: array, items: {type: object, required: [type, message], properties: {type: {type: string, enum: [snapshot_mismatch, provider_unconfirmed, ledger_payout_mismatch, payout_failed, employment_payroll]}, message: {type: string}, assignmentId: {type: string, format: uuid}, payoutId: {type: string, format: uuid}, amountYen: {type: integer}}}}
...